requirements.txt
playwright==1.49.0
requests==2.31.0
beautifulsoup4==4.12.3
pyyaml==6.0.1
watchdog==4.0.0
schedule==1.2.1
win10toast==0.9
